[TCFMM-17] Extend geometric mean oracle
Problem: the article says that we need to store sum of `i_c` over all
the seconds during which the contract existed. This is necessary to
provide Geometric Mean oracle.

Moreover, we have to remember this sum not only for one recent moment of
time as it is done now, but for a series of snapshots with sums.
Initially we can allow only one snapshot, but this number should be
increase-able using `increase_observation_count` entrypoint.

Solution: implement this.

Also, we update the view. Now the user can request a value for the most
recent available timestamp as before, or for a specific timestamp if it
is within the range of yet stored values.

Some implementation details below.

Buffer for storing cummulative time-weighted tick indices:

For Uniswap V3 we want to keep cummulative time-weighted tick indices
for the last several timestamps.

We store these values along with the respective timestamps in a
nat-indexed map. A new value is added each some action is performed in
the contract, but not more often than once per block.
Note that we cannot use `timestamp` as a key in the map, because then to
access a value we would need to know in advance at which timestamps
the values are registered.

We limit the number of values kept for the sake of avoiding paying extra
storage costs. Anyone can ask for increasing the number of values kept,
but he will also pay for the new slots in the storage.
So when an increase of slots number is demanded, we add dummy values to
the map. Over time, when recording new values, as long as we overwrite
those dummy values we don't erase the oldest values (like in the Snake
game when the snake consumes some food).

In Tezos, when a transaction adds something to a `big_map` and
simultaneously erases another element of the same size in that
`big_map`, no storage costs are payed. Thanks to this, we don't need to
implement anything like ring buffer, we can just use `big_map` and use
ever-increasing indices as keys.

Searching a value by timestamp:

We do not store cummulative time-weighted tick indices for every
possible timestamp, but only remember the new value when the contract is
called. This means that there can be arbitrary periods between two
values being recorded.

We still want to allow the external contract to ask a value at an
arbitrary timestamp. Here our computation relies on two facts:
* In our buffer, timestamps increase strictly monotonically.
* Between any two adjacent records in that buffer, the value in demand
  grows linearly (because between any two records, `cur_tick_index`
  stays constant).

So we use binary search + interpolation to evaluate the value at any
given timestamp (if it is within the remembered range).